+++ planner/cardinality/BUILD.bazel
@@ -1,0 +1,70 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "cardinality",
+    srcs = [
+        "pseudo.go",
+        "selectivity.go",
+        "trace.go",
+    ],
+    importpath = "github.com/pingcap/tidb/planner/cardinality",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//expression",
+        "//parser/ast",
+        "//parser/format",
+        "//parser/mysql",
+        "//planner/util",
+        "//planner/util/debugtrace",
+        "//sessionctx",
+        "//statistics",
+        "//types",
+        "//types/parser_driver",
+        "//util/chunk",
+        "//util/logutil",
+        "//util/ranger",
+        "//util/tracing",
+        "@com_github_pingcap_errors//:errors",
+        "@org_golang_x_exp//maps",
+        "@org_uber_go_zap//:zap",
+    ],
+)
+
+go_test(
+    name = "cardinality_test",
+    timeout = "short",
+    srcs = [
+        "main_test.go",
+        "selectivity_test.go",
+    ],
+    data = glob(["testdata/**"]),
+    embed = [":cardinality"],
+    flaky = True,
+    shard_count = 23,
+    deps = [
+        "//config",
+        "//domain",
+        "//kv",
+        "//parser/model",
+        "//parser/mysql",
+        "//planner/core",
+        "//session",
+        "//sessionctx",
+        "//sessionctx/stmtctx",
+        "//sessionctx/variable",
+        "//statistics",
+        "//statistics/handle",
+        "//testkit",
+        "//testkit/testdata",
+        "//testkit/testmain",
+        "//testkit/testsetup",
+        "//types",
+        "//util/codec",
+        "//util/collate",
+        "//util/mock",
+        "//util/ranger",
+        "@com_github_pingcap_failpoint//:failpoint",
+        "@com_github_stretchr_testify//require",
+        "@org_uber_go_goleak//:goleak",
+    ],
+)
