+++ planner/cardinality/pseudo.go
@@ -1,0 +1,82 @@
+// Copyright 2023 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package cardinality
+
+import (
+	"math"
+
+	"github.com/pingcap/tidb/expression"
+	"github.com/pingcap/tidb/parser/ast"
+	"github.com/pingcap/tidb/parser/mysql"
+	"github.com/pingcap/tidb/statistics"
+)
+
+const (
+	pseudoEqualRate = 1000
+	pseudoLessRate  = 3
+)
+
+// If one condition can't be calculated, we will assume that the selectivity of this condition is 0.8.
+const selectionFactor = 0.8
+
+func pseudoSelectivity(coll *statistics.HistColl, exprs []expression.Expression) float64 {
+	minFactor := selectionFactor
+	colExists := make(map[string]bool)
+	for _, expr := range exprs {
+		fun, ok := expr.(*expression.ScalarFunction)
+		if !ok {
+			continue
+		}
+		colID := getConstantColumnID(fun.GetArgs())
+		if colID == unknownColumnID {
+			continue
+		}
+		switch fun.FuncName.L {
+		case ast.EQ, ast.NullEQ, ast.In:
+			minFactor = math.Min(minFactor, 1.0/pseudoEqualRate)
+			col, ok := coll.Columns[colID]
+			if !ok {
+				continue
+			}
+			colExists[col.Info.Name.L] = true
+			if mysql.HasUniKeyFlag(col.Info.GetFlag()) {
+				return 1.0 / float64(coll.RealtimeCount)
+			}
+		case ast.GE, ast.GT, ast.LE, ast.LT:
+			minFactor = math.Min(minFactor, 1.0/pseudoLessRate)
+			// FIXME: To resolve the between case.
+		}
+	}
+	if len(colExists) == 0 {
+		return minFactor
+	}
+	// use the unique key info
+	for _, idx := range coll.Indices {
+		if !idx.Info.Unique {
+			continue
+		}
+		unique := true
+		for _, col := range idx.Info.Columns {
+			if !colExists[col.Name.L] {
+				unique = false
+				break
+			}
+		}
+		if unique {
+			return 1.0 / float64(coll.RealtimeCount)
+		}
+	}
+	return minFactor
+}
