+++ statistics/BUILD.bazel
@@ -29,18 +29,15 @@ go_library(
         "//kv",
         "//parser/ast",
         "//parser/charset",
-        "//parser/format",
         "//parser/model",
         "//parser/mysql",
         "//parser/terror",
-        "//planner/util",
         "//planner/util/debugtrace",
         "//sessionctx",
         "//sessionctx/stmtctx",
         "//sessionctx/variable",
         "//tablecodec",
         "//types",
-        "//types/parser_driver",
         "//util/chunk",
         "//util/codec",
         "//util/collate",
@@ -75,30 +72,25 @@ go_test(
         "main_test.go",
         "sample_test.go",
         "scalar_test.go",
-        "selectivity_test.go",
         "statistics_test.go",
         "trace_test.go",
     ],
     data = glob(["testdata/**"]),
     embed = [":statistics"],
     flaky = True,
-    shard_count = 50,
+    shard_count = 48,
     deps = [
         "//config",
-        "//domain",
         "//executor",
         "//expression",
         "//infoschema",
-        "//kv",
         "//parser",
         "//parser/ast",
         "//parser/model",
         "//parser/mysql",
         "//planner/core",
-        "//session",
         "//sessionctx",
         "//sessionctx/stmtctx",
-        "//sessionctx/variable",
         "//statistics/handle",
         "//testkit",
         "//testkit/testdata",
